refactored all game phases into components, accessing app context
